Refactor user management: update UserEntity and UserApiModel to use Long for IDs, modify UserDao and repositories accordingly, and implement User mappers for data transformation

+++ StudyApp/app/src/main/java/com/example/studyapp/data/local/dao/UserDao.kt
@@ -1,7 +1,6 @@
 package com.example.studyapp.data.local.dao
 
 import androidx.room.Dao
-import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
@@ -23,5 +22,5 @@ interface UserDao {
 //    @Delete
 //    suspend fun deleteUser(user: UserEntity)
     @Query("DELETE FROM users WHERE id = :userId")
-    suspend fun deleteUser(userId: Int)
+    suspend fun deleteUser(userId: Long)
 }
